use component for list and editor

+++ client/src/pages/admin/NewsDashboard/NewsList/index.jsx
@@ -1,64 +1,42 @@
 import { useDispatch, useSelector } from "react-redux";
-import { deleteNews, getNews } from "../newsSlice";
-import Button from "components/Button";
-import style from "./NewsList.module.scss";
+import { getNews, deleteNews } from "../newsSlice";
+import { newsAPI } from "api";
+import { deleteObject, ref } from "firebase/storage";
+import { storage } from "config/firebase";
+import DataTable from "components/DataTable";
 
 const NewsList = ({ setId }) => {
   const dispatch = useDispatch();
   const news = useSelector((state) => state.news);
 
+  const deleteHandle = (id) => {
+    if (window.confirm("Are you sure to delete this course")) {
+      newsAPI.getOneNews(id).then((res) => {
+        const currentRef = ref(storage, res.data.thumbnail);
+        deleteObject(currentRef);
+        res.data.images &&
+          res.data.images.forEach((image) => {
+            const currentRef = ref(storage, image);
+            deleteObject(currentRef);
+          });
+        dispatch(deleteNews(id));
+      });
+    }
+  };
   const loadHandle = () => {
     dispatch(getNews({ skip: news.length, limit: 5 }));
   };
-  const deleteHandle = (id) => {
-    dispatch(deleteNews(id));
-  };
 
   return (
-    <div className={style.newsList}>
-      <table>
-        <thead>
-          <tr>
-            <th>Thumbnail</th>
-            <th>Title</th>
-            <th>Date</th>
-            <th>
-              <div className="d-flex justify-content-center">
-                <Button variant="outline" onClick={() => setId(0)}>
-                  Add New News
-                </Button>
-              </div>
-            </th>
-          </tr>
-        </thead>
-        <tbody>
-          {news.map((news) => (
-            <tr key={news._id}>
-              <td>
-                <img src={news.thumbnail} alt="" />
-              </td>
-              <td>
-                <span>{news.title}</span>
-              </td>
-              <td>
-                <span>{news.updatedAt.split("T")[0]}</span>
-              </td>
-              <td>
-                <div className="d-flex justify-content-center">
-                  <Button variant="outline" onClick={() => setId(news._id)}>
-                    Edit
-                  </Button>
-                  <Button variant="outline" onClick={() => deleteHandle(news._id)}>
-                    Delete
-                  </Button>
-                </div>
-              </td>
-            </tr>
-          ))}
-        </tbody>
-      </table>
-      <Button variant="outline" className={style.loadButton} onClick={loadHandle}>Load more</Button>
-    </div>
+    <DataTable
+      fields={["thumbnail", "title", "date"]}
+      data={news.map((item) => {
+        const formatDate = { ...item };
+        formatDate.date = formatDate.createdAt.split("T")[0];
+        return formatDate;
+      })}
+      actionHandler={[setId, deleteHandle, loadHandle]}
+    />
   );
 };
